+++ lgtv-drivers/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.1] - 2025-01-25
+### Fixed
+- Improve the detection of TV startup using a fast-ping mechanism - `@hubitrep`
+- Fix the "Start Video" and "Speak" commands for webOS 5 - `@hubitrep`
+
 ## [1.5.0] - 2025-01-25
 ### Added
 - Add "SpeechSynthesis/Speak" command for TTS announcements